[PyCRAP] lower case name to prevent conflicts with classes.
